refactored some components and improved items ui
