+++ src/components/Map.js
@@ -1,7 +1,6 @@
 import React from "react";
 import L from "leaflet";
 import { MapContainer, TileLayer, Marker, Popup } from "react-leaflet";
-import './Map.css'
 
 var marker = L.icon({
     iconUrl: './MapPin.webp',
